+++ CHANGELOG.md
@@ -1481,3 +1481,4 @@
 - [2026-05-26 21:36] fix(security): prevent double tap duplicate action in security
 - [2026-05-26 21:43] test(theme): implement widget integration tests for theme
 - [2026-05-26 22:34] feat(network): add bi-directional event bus handler for network
+- [2026-05-26 22:37] fix(perf): fix memory leak by cancelling perf subscription